+++ src/core/platforms/ctrader/positions/CTraderPositionParameters.ts
@@ -1,6 +1,7 @@
-import { MidaPositionParameters, } from "@reiryoku/mida";
+import { MidaEmitter, MidaPositionParameters, } from "@reiryoku/mida";
 import { CTraderConnection, } from "@reiryoku/ctrader-layer";
 
 export type CTraderPositionParameters = MidaPositionParameters & {
     connection: CTraderConnection;
+    cTraderEmitter: MidaEmitter;
 };